Move kotlin coroutines instrumentation to top-level instrumentation module
