+++ instrumentation/java-concurrent/java-concurrent.gradle
@@ -8,13 +8,6 @@ apply from: "$rootDir/gradle/instrumentation.gradle"
 
 muzzle {
   pass {
-    extraDependency('org.jetbrains.kotlin:kotlin-stdlib-common:1.3.72')
-    extraDependency('org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.7')
     coreJdk()
   }
-}
-dependencies {
-  // This is needed for kotlin coroutine suspend/resume instrumentation
-  compileOnly 'org.jetbrains.kotlin:kotlin-stdlib-common:1.3.72'
-  compileOnly 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.7'
 }
