+++ settings.gradle
@@ -97,7 +97,6 @@ include ':instrumentation:java-classloader:jboss-testing'
 include ':instrumentation:java-classloader:osgi-testing'
 include ':instrumentation:java-classloader:tomcat-testing'
 include ':instrumentation:java-concurrent'
-include ':instrumentation:java-concurrent:kotlin-testing'
 include ':instrumentation:jaxrs:jaxrs-1.0'
 include ':instrumentation:jaxrs:jaxrs-2.0'
 include ':instrumentation:jaxrs:jaxrs-2.0:jaxrs-2.0-jersey-2.0'
@@ -116,6 +115,8 @@ include ':instrumentation:jsp-2.3'
 include ':instrumentation:kafka-clients-0.11'
 include ':instrumentation:kafka-streams-0.11'
 include ':instrumentation:khttp-0.1'
+include ':instrumentation:kotlin-coroutines'
+include ':instrumentation:kotlin-coroutines:kotlin-testing'
 include ':instrumentation:kubernetes-client-7.0'
 include ':instrumentation:lettuce:lettuce-4.0'
 include ':instrumentation:lettuce:lettuce-5.0'
